fixed bug of controlls not working when loading from "office scene" -> deleted controlschemes from playercontrolls.inputactions
delete schemes !!! -> So in effect, the intention for having two control schemes, one for keyboard and one for gamepad, would be that keyboard input is not meant to be used concurrently with gamepad input. While the keyboard bindings are active, the gamepad bindings will not be. If both are in the same control scheme, however, they will concurrently be active side by side. This means that, for example, if the player has both a keyboard and a gamepad connected, the system will establish a connection to both even if the player only uses one of them.

To conclude this and make it slightly more confusing, though, worth mentioning that most of the APIs do not respect control schemes except when being told to. PlayerInput is different. If there are control schemes present in the asset, it will use and respect them. However, most of the other, lower level APIs will blindly operate on all bindings present in the asset except if restricted to specific ones.